# excourt_re/excourt_re-master/system/app/static/swagger.yaml
openapi: 3.0.0
info:
  title: ExCourt API
  description: API for court management system
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /friend/add:
    post:
      summary: Add a friend
      tags:
        - Friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - my_id
                - search_id
              properties:
                my_id:
                  type: integer
                  description: ID of the user adding a friend
                search_id:
                  type: integer
                  description: ID of the user to be added as friend
      responses:
        '201':
          description: Friend successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Friend added successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
  /friend/delete:
    post:
      summary: Delete a friend
      tags:
        - Friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - my_id
                - search_id
              properties:
                my_id:
                  type: integer
                  description: ID of the user deleting a friend
                search_id:
                  type: integer
                  description: ID of the user to be removed from friends
      responses:
        '200':
          description: Friend successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Friend deleted successfully
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Missing required parameters
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
  /friend/getall:
    post:
      summary: Get list of all friends
      tags:
        - Friends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student_id
              properties:
                student_id:
                  type: integer
                  description: Student ID
      responses:
        '200':
          description: Friends list successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Friends list retrieved successfully
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        student_id:
                          type: integer
                          description: Student ID
                        name:
                          type: string
                          description: Student name
                        profile_url:
                          type: string
                          description: Student avatar URL
                        nickname:
                          type: string
                          description: Student nickname
        '400':
          description: Missing student ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Missing required parameters
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] 